fix: import/export of modules [MONET-1338]
