+++ packages/ecommerce-app-base/jest-setup.ts
@@ -1,5 +1,6 @@
 import '@testing-library/jest-dom';
 import { configure } from '@testing-library/react';
+
 configure({
   testIdAttribute: 'data-test-id',
 });
