+++ packages/ecommerce-app-base/src/Editor/index.tsx
@@ -1,0 +1,2 @@
+export { Field } from './Field';
+export { IntegrationProvider, useIntegration } from './IntegrationContext';
