+++ packages/ecommerce-app-base/src/SkuPicker/ProductSelectionList/OtherProductsCounter.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import tokens from '@contentful/f36-tokens';
 import { css } from 'emotion';
