+++ packages/ecommerce-app-base/src/SkuPicker/ProductSelectionList/ProductSelectionList.spec.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import { render, cleanup } from '@testing-library/react';
 import { Props, ProductSelectionList } from '.';
 import productPreviews from '../../__mocks__/productPreviews';
