+++ packages/ecommerce-app-base/src/SkuPicker/index.tsx
@@ -1,0 +1,1 @@
+export { SkuPicker } from './SkuPicker';
